requirements.txt
apache-airflow-providers-amazon==8.15.0